Add Paginates, Update Dashboard teacher user, Update Catatan all v1, dan Update All Biodata v2

+++ app/Http/Controllers/JurusanController.php
@@ -4,17 +4,25 @@ namespace App\Http\Controllers;
 
 use App\Models\Jurusan;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use APp\Enums\UserRole;
 
 class JurusanController extends Controller
 {
     public function index()
     {
-        $jurusans = Jurusan::all();
+        if (Auth::user()->role !== UserRole::Teacher) {
+            abort(403, 'Unauthorized action.');
+        }
+        $jurusans = Jurusan::paginate(5);
         return view('jurusans.index', compact('jurusans'));
     }
 
     public function create()
     {
+        if (Auth::user()->role !== UserRole::Teacher) {
+            abort(403, 'Unauthorized action.');
+        }
         return view('jurusans.create');
     }
 
@@ -31,6 +39,9 @@ class JurusanController extends Controller
 
     public function edit(Jurusan $jurusan)
     {
+        if (Auth::user()->role !== UserRole::Teacher) {
+            abort(403, 'Unauthorized action.');
+        }
         return view('jurusans.edit', compact('jurusan'));
     }
 
@@ -47,12 +58,18 @@ class JurusanController extends Controller
 
     public function show(Jurusan $jurusan)
     {
+        if (Auth::user()->role !== UserRole::Teacher) {
+            abort(403, 'Unauthorized action.');
+        }
         $rooms = $jurusan->rooms; // Ambil semua room yang terkait dengan jurusan
         return view('jurusans.show', compact('jurusan', 'rooms'));
     }
 
     public function destroy(Jurusan $jurusan)
     {
+        if (Auth::user()->role !== UserRole::Teacher) {
+            abort(403, 'Unauthorized action.');
+        }
         $jurusan->delete();
 
         return redirect()->route('jurusans.index')->with('success', 'Jurusan deleted successfully.');
